moved and refactored lookUpInMaster()
since BaseDataPipe was the only object using this function, decided to move the function inside BaseDataPipe. Changed the mysql functions to use the DBManager PDO object. Includes a test for a positive match on the supplied object ID and a negative match. The object ID used should be included in the sql/projectfileInserts.sql.